+++ rollup.config.js
@@ -1,11 +1,10 @@
-import typescript from 'rollup-plugin-typescript2'
-import commonjs from 'rollup-plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import resolve from 'rollup-plugin-node-resolve'
-import url from 'rollup-plugin-url'
-import svgr from '@svgr/rollup'
+import typescript from 'rollup-plugin-typescript2';
+import commonjs from 'rollup-plugin-commonjs';
+import external from 'rollup-plugin-peer-deps-external';
+import resolve from 'rollup-plugin-node-resolve';
+import url from 'rollup-plugin-url';
 
-import pkg from './package.json'
+import pkg from './package.json';
 
 export default {
   input: 'src/index.tsx',
@@ -23,15 +22,15 @@ export default {
       sourcemap: true,
     },
   ],
+  external: ['stream'],
   plugins: [
     external(),
     url(),
-    svgr(),
-    resolve(),
+    resolve({ preferBuiltins: true }),
     typescript({
       rollupCommonJSResolveHack: true,
       clean: true,
     }),
     commonjs(),
   ],
-}
+};